requirements.txt
# Core Backend Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Audio Processing
pydub==0.25.1
faster-whisper==0.10.0

# NLP & ML
spacy==3.7.2
transformers==4.37.0
torch>=2.2.0
negspacy==1.0.4

# Database
sqlalchemy==2.0.25
alembic==1.13.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Utilities
python-dateutil==2.8.2
email-validator==2.3.0

# PostgreSQL (for production)
psycopg2-binary==2.9.9
